Create README.md
An opinion poll also referred to as a poll is a research methodology carried out by an individual or group of individuals in order to gather the opinions of a sample of respondents from a population. Opinion polls are constructed to show the responses or opinions of a population (respondents) through the use of a questionnaire then deduce their responses in ratio or within confidence intervals. Polls is a mobile based polling application that enables an instructor to conduct polling in class and instantly assess student comprehension with live, in-class polling. You can create opinion polls with no set answer or create a poll question that requires a correct response. You can also stop the poll at any time and show and hide poll results with students. Students can also view and take polls available to them. Poll share organized as current or closed; current polls are sorted as to whether or not they have been answered, if an instructor allows, students can also view results for a closed poll.